+++ NEWS.md
@@ -2,6 +2,7 @@
 * Minor bugfix that let to an error when running a model without covariates.
 * `extract_draws()` now accepts several parameters at once.
 * Minor documentation fixes
+* Bugfix: Removing order constrained from adjacent category model.
 
 # bvarnet 1.0.2